requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# AI Providers
anthropic>=0.40.0         # Claude - REQUIRED
openai>=1.50.0            # OpenAI - optional alternative

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# HTTP Client
httpx>=0.28.0

# Scheduling
apscheduler>=3.10.4
pytz>=2024.1

# PDF Generation
reportlab>=4.2.0

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.9.0
